Revert ""Updated AuthController, DashboardController, and layouts/components to refactor authentication and navigation logic""
This reverts commit ff17f349e4ad5470466bdf365860d4fcf2cfaba5.
