+++ app/Http/Controllers/AuthController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Inertia\Inertia;
 
 class AuthController extends Controller
 {
@@ -22,7 +21,7 @@ class AuthController extends Controller
     public function signIn(Request $request)
     {
         // Validate input fields
-        $request->validate(rules: [
+        $request->validate([
             'emailOrStudentNumber' => 'required',
             'password' => 'required',
         ]);
@@ -43,7 +42,7 @@ class AuthController extends Controller
     
         // If the input format is invalid, return error
         if (!$credentials) {
-            return back()->withErrors(['emailOrStudentNumber' => 'Invalid email or student number format.']);
+            return back()->withErrors(['Email or Student Number' => 'Invalid email or student number format.']);
         }
     
         // Attempt to authenticate with the determined credentials
@@ -52,7 +51,7 @@ class AuthController extends Controller
         }
     
         // Return error for invalid credentials
-        return back()->withErrors(['emailOrStudentNumber' => 'Invalid Credentials.']);
+        return back()->withErrors(['Email or Student Number' => 'Invalid Credentials.']);
     }
     
     
@@ -82,8 +81,7 @@ class AuthController extends Controller
     public function logout()
     {
         Auth::logout();
-        sleep(2);
-        return redirect('/');
+        return inertia('Authentication/Login');
     }
 
     /**
