--- resources/js/Pages/ProgramHead/AddQuestion.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AddQuestion = () => {
-    return <div className="">AddQuestion</div>;
-};
-
-export default AddQuestion;
