+++ resources/js/components/AuthComponents/AuthForm.jsx
@@ -70,18 +70,14 @@ const AuthForm = ({
                             value={data[name]}
                             onChange={(e) => setData(name, e.target.value)}
                             className={`input input-bordered bg-white w-full border-[#9BB1AF] ${
-                                errors[name] || serverErrors[name]
-                                    ? "border-red-500"
-                                    : ""
+                                errors[name] ? "border-red-500" : ""
                             }`}
                         />
-                        {/* Client-side validation errors */}
                         {errors[name] && (
                             <span className="text-red-500 text-sm">
                                 {errors[name]}
                             </span>
                         )}
-                        {/* Server-side validation errors */}
                         {serverErrors[name] && (
                             <span className="text-red-500 text-sm">
                                 {serverErrors[name]}
