+++ resources/js/components/ProgramHead/Sidebar.jsx
@@ -1,9 +1,14 @@
 import React, { useState } from "react";
 import { TbCloudQuestion } from "react-icons/tb";
 
-const Sidebar = ({ isOpen, toggleSidebar }) => {
+const Sidebar = () => {
+    const [isOpen, setIsOpen] = useState(false); // Sidebar is hidden by default on mobile
     const [activeSummary, setActiveSummary] = useState(null); // Track active summary
     const [activeItem, setActiveItem] = useState(null); // Track active list item
+
+    const toggleSidebar = () => {
+        setIsOpen(!isOpen);
+    };
 
     // Function to handle clicking an item
     const handleItemClick = (summaryIndex, itemIndex) => {
@@ -14,9 +19,9 @@ const Sidebar = ({ isOpen, toggleSidebar }) => {
     return (
         <>
             <aside
-                className={`fixed left-0 top-0 flex h-screen flex-col duration-300 ease-linear lg:static lg:translate-x-0  bg-[#42604C] text-white z-40 ${
-                    isOpen ? "translate-x-0 w-72" : "-translate-x-full w-16"
-                }`}
+                className={`absolute left-0 top-0 flex h-screen flex-col duration-300 ease-linear lg:static lg:translate-x-0 -translate-x-full bg-[#42604C] text-white ${
+                    isOpen ? "w-72" : "w-16"
+                } ${isOpen ? "block" : "hidden"} md:block`}
             >
                 <div
                     className={`flex items-center justify-between gap-2 py-6 ${
@@ -105,7 +110,7 @@ const Sidebar = ({ isOpen, toggleSidebar }) => {
                                         </li>
                                     </ul>
                                 </details>
-                                <details>
+                                <details close>
                                     <summary
                                         className={`hover:bg-green-100 hover:text-black ${
                                             activeSummary === 2
@@ -160,6 +165,28 @@ const Sidebar = ({ isOpen, toggleSidebar }) => {
                     </nav>
                 </div>
             </aside>
+
+            {/* Burger Icon (only visible on mobile) */}
+            <button
+                className={`btn btn-square btn-ghost fixed top-4 left-4 z-50 md:hidden ${
+                    isOpen ? "hidden" : "flex"
+                }`}
+                onClick={toggleSidebar}
+            >
+                <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    className="inline-block w-6 h-6 stroke-current"
+                >
+                    <path
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                        strokeWidth="2"
+                        d="M4 6h16M4 12h16M4 18h16"
+                    />
+                </svg>
+            </button>
         </>
     );
 };
